+++ truffle/src/com.oracle.truffle.runtime/src/com/oracle/truffle/runtime/OptimizedCallTarget.java
@@ -295,6 +295,7 @@ public abstract class OptimizedCallTarget implements TruffleCompilable, RootCall
      * after the compilation has started does not reset the task until the compilation finishes.
      */
     private volatile CompilationTask compilationTask;
+    private int successfulCompilationsCount;
 
     private volatile boolean needsSplit;
 
@@ -934,6 +935,15 @@ public abstract class OptimizedCallTarget implements TruffleCompilable, RootCall
 
                     try {
                         assert compilationTask == null;
+                        if (engine.maximumCompilations >= 0 && successfulCompilationsCount >= engine.maximumCompilations) {
+                            compilationFailed = true;
+                            runtime().getListener().onCompilationStarted(this, new PresubmitFailureCompilationTask(engine.firstTierOnly, lastTier));
+                            String failureReason = String.format("Maximum compilation count %d reached.", engine.maximumCompilations);
+                            runtime().getListener().onCompilationFailed(this, failureReason, true, true,
+                                            lastTier ? 2 : 1, null);
+                            handleCompilationFailure(() -> failureReason, false, true, true);
+                            return false;
+                        }
                         this.compilationTask = task = runtime().submitForCompilation(this, lastTier);
                     } catch (RejectedExecutionException e) {
                         return false;
@@ -1092,6 +1102,11 @@ public abstract class OptimizedCallTarget implements TruffleCompilable, RootCall
     }
 
     @Override
+    public void onCompilationSuccess(int compilationTier, boolean lastTier) {
+        successfulCompilationsCount++;
+    }
+
+    @Override
     public final boolean onInvalidate(Object source, CharSequence reason, boolean wasActive) {
         cachedNonTrivialNodeCount = -1;
         if (wasActive) {
@@ -1109,6 +1124,10 @@ public abstract class OptimizedCallTarget implements TruffleCompilable, RootCall
             }
         }
 
+        handleCompilationFailure(serializedException, silent, bailout, permanentBailout);
+    }
+
+    private void handleCompilationFailure(Supplier<String> serializedException, boolean silent, boolean bailout, boolean permanentBailout) {
         ExceptionAction action;
         if (bailout && !permanentBailout) {
             /*
@@ -1906,4 +1925,31 @@ public abstract class OptimizedCallTarget implements TruffleCompilable, RootCall
         return engine.getCompilerOptions();
     }
 
+    private static class PresubmitFailureCompilationTask extends AbstractCompilationTask {
+        private final boolean firstTierOnly;
+        private final boolean lastTier;
+
+        PresubmitFailureCompilationTask(boolean firstTierOnly, boolean lastTier) {
+            this.firstTierOnly = firstTierOnly;
+            this.lastTier = lastTier;
+        }
+
+        @Override
+        public boolean isCancelled() {
+            return false;
+        }
+
+        @Override
+        public boolean isLastTier() {
+            return lastTier;
+        }
+
+        @Override
+        public boolean hasNextTier() {
+            if (lastTier) {
+                return false;
+            }
+            return !firstTierOnly;
+        }
+    }
 }